"Added useCallback and useFocusEffect to Groups screen, implemented createGroup function in storage, and updated NewGroup screen to use createGroup function"

+++ src/storage/group/createGroup.ts
@@ -1,11 +1,17 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 import { getAllGroups } from "./getAllGroups";
+import { AppError } from "@utils/AppError";
 
-export async function groupCreate(newGroup: string) {
+export async function createGroup(newGroup: string) {
   try {
-    await AsyncStorage.setItem(GROUP_COLLECTION, newGroup);
     const storedGroups = await getAllGroups();
+
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
+    }
 
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
